--- src/Common/src/Internal/Runtime/EEType.cs
@@ -1,56 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-using System.Diagnostics;
-
-namespace Internal.Runtime
-{
-    internal partial class EEType
-    {
-        // These masks and paddings have been chosen so that the ValueTypePadding field can always fit in a byte of data
-        // if the alignment is 8 bytes or less. If the alignment is higher then there may be a need for more bits to hold
-        // the rest of the padding data.
-        // If paddings of greater than 7 bytes are necessary, then the high bits of the field represent that padding
-        private const UInt32 ValueTypePaddingLowMask = 0x7;
-        private const UInt32 ValueTypePaddingHighMask = 0xFFFFFF00;
-        private const UInt32 ValueTypePaddingMax = 0x07FFFFFF;
-        private const int ValueTypePaddingHighShift = 8;
-        private const UInt32 ValueTypePaddingAlignmentMask = 0xF8;
-        private const int ValueTypePaddingAlignmentShift = 3;
-
-        /// <summary>
-        /// Compute the encoded value type padding and alignment that are stored as optional fields on an
-        /// <c>EEType</c>. This padding as added to naturally align value types when laid out as fields
-        /// of objects on the GCHeap. The amount of padding is recorded to allow unboxing to locals /
-        /// arrays of value types which don't need it.
-        /// </summary>
-        internal static UInt32 ComputeValueTypeFieldPaddingFieldValue(UInt32 padding, UInt32 alignment)
-        {
-            // For the default case, return 0
-            if ((padding == 0) && (alignment == IntPtr.Size))
-                return 0;
-
-            UInt32 alignmentLog2 = 0;
-            Debug.Assert(alignment != 0);
-
-            while ((alignment & 1) == 0)
-            {
-                alignmentLog2++;
-                alignment = alignment >> 1;
-            }
-            Debug.Assert(alignment == 1);
-
-            Debug.Assert(ValueTypePaddingMax >= padding);
-
-            // Our alignment values here are adjusted by one to allow for a default of 0 (which represents pointer alignment)
-            alignmentLog2++;
-
-            UInt32 paddingLowBits = padding & ValueTypePaddingLowMask;
-            UInt32 paddingHighBits = ((padding & ~ValueTypePaddingLowMask) >> ValueTypePaddingAlignmentShift) << ValueTypePaddingHighShift;
-            UInt32 alignmentLog2Bits = alignmentLog2 << ValueTypePaddingAlignmentShift;
-            Debug.Assert((alignmentLog2Bits & ~ValueTypePaddingAlignmentMask) == 0);
-            return paddingLowBits | paddingHighBits | alignmentLog2Bits;
-        }
-    }
-}
